+++ core/src/main/scala/astraea/spark/rasterframes/RasterFunctions.scala
@@ -31,6 +31,7 @@ import astraea.spark.rasterframes.{functions => F}
 import geotrellis.proj4.CRS
 import geotrellis.raster.mapalgebra.local.LocalTileBinaryOp
 import geotrellis.raster.{CellType, Tile}
+import geotrellis.vector.Extent
 import org.apache.spark.annotation.Experimental
 import org.apache.spark.sql._
 import org.apache.spark.sql.functions._
@@ -60,7 +61,11 @@ trait RasterFunctions {
   def tile_dimensions(col: Column): Column = GetDimensions(col)
 
   /** Extracts the bounding box of a geometry as a JTS envelope. */
+  @deprecated("Replace usages of this with `st_extent`", "11/4/2018")
   def envelope(col: Column): TypedColumn[Any, Envelope] = GetEnvelope(col)
+
+  /** Extracts the bounding box of a geometry as an Extent */
+  def st_extent(col: Column): TypedColumn[Any, Extent] = GeometryToExtent(col)
 
   /** Flattens Tile into a double array. */
   def tile_to_array_double(col: Column): TypedColumn[Any, Array[Double]] =