+++ core/src/main/scala/astraea/spark/rasterframes/expressions/package.scala
@@ -66,6 +66,7 @@ package object expressions {
     registry.registerExpression[SetCellType]("rf_convert_cell_type")
     registry.registerExpression[GetDimensions]("rf_tile_dimensions")
     registry.registerExpression[BoundsToGeometry]("rf_bounds_geometry")
+    registry.registerExpression[GeometryToExtent]("st_extent")
     registry.registerExpression[Subtract]("rf_local_subtract")
     registry.registerExpression[Multiply]("rf_local_multiply")
     registry.registerExpression[Divide]("rf_local_divide")